PL-1950 Fix GeoSlicerBase build Windows pipeline

Approved-by: Rodrigo Schwedersky

+++ tools/build_and_pack.py
@@ -20,8 +20,7 @@ logger.addHandler(logging.FileHandler(filename=log_file_path.as_posix()))
 logger.setLevel(logging.INFO)
 
 
-
-def get_working_dir():
+def get_working_dir(args):
     if args.avoid_long_path and sys.platform == "win32":
         path = Path(Path.home().drive + "/gsb")
     else:
@@ -30,10 +29,21 @@ def get_working_dir():
     return path
 
 
-def generate_project_command(args):
-    output_directory_path = get_working_dir()
+def get_project_build_command(args):
+    output_directory_path = get_working_dir(args)
     if sys.platform == "win32":
-        command = f"cmake -S {args.source.as_posix()} -B {output_directory_path.as_posix()} -DCMAKE_BUILD_TYPE:STRING={args.type} -DSlicer_SKIP_ROOT_DIR_MAX_LENGTH_CHECK:BOOL=TRUE"
+        command = [
+            "cmake",
+            "-G",
+            "Visual Studio 16 2019",
+            "-A",
+            "x64",
+            "-DQt5_DIR:PATH=C:/Qt/5.15.2/msvc2019_64/lib/cmake/Qt5",
+            "-S",
+            args.source.as_posix(),
+            "-B",
+            output_directory_path.as_posix(),
+        ]
     else:
         command = f"cmake -S {args.source.as_posix()} -B {output_directory_path.as_posix()} -DCMAKE_BUILD_TYPE:STRING={args.type}"
 
@@ -42,14 +52,17 @@ def generate_project_command(args):
 
 def build_command(args):
     if sys.platform == "win32":
-        command = f"msbuild GeoSlicer.sln /p:Configuration={args.type} /m:{args.jobs}"
+        output_directory_path = get_working_dir(args)
+        command = (
+            f"cmake --build {output_directory_path.as_posix()} --config {args.type}"
+        )
     else:
         command = f"make -j{args.jobs}"
 
     return command
 
 
-def generate_package_command(args):
+def get_package_command(args):
     if sys.platform == "win32":
         command = f"msbuild PACKAGE.vcxproj /p:Configuration={args.type} /m:{args.jobs}"
     else:
@@ -59,7 +72,11 @@ def generate_package_command(args):
 
 
 def process(args):
-    output_directory_path = get_working_dir()
+    if args.only_export:
+        only_export_process(args)
+        return
+
+    output_directory_path = get_working_dir(args)
     if args.no_cache and output_directory_path.exists():
         logger.info("Deleting old build directory remainings")
         shutil.rmtree(output_directory_path, onerror=onerror)
@@ -70,13 +87,15 @@ def process(args):
 
     # Run cmake to generate buildable project
     logger.info("Generating buildable project...")
-    command_as_list = generate_project_command(args).split()
+    command_as_list = get_project_build_command(args)
     run_subprocess(command_as_list, shell=subprocess_as_shell)
 
     # Build project
     logger.info("Building project...")
     command_as_list = build_command(args).split()
-    run_subprocess(command_as_list, shell=subprocess_as_shell, cwd=output_directory_path.as_posix())
+    run_subprocess(
+        command_as_list, shell=subprocess_as_shell, cwd=output_directory_path.as_posix()
+    )
 
     # Pack the application
     logger.info("Packaging application...")
@@ -84,17 +103,23 @@ def process(args):
     if not slicer_build_file_path.exists():
         raise RuntimeError("Slicer-build folder not found in the project directory.")
 
-    command_as_list = generate_package_command(args).split()
-    run_subprocess(command_as_list, shell=subprocess_as_shell, cwd=slicer_build_file_path.as_posix())
+    command_as_list = get_package_command(args)
+    run_subprocess(
+        command_as_list,
+        shell=subprocess_as_shell,
+        cwd=slicer_build_file_path.as_posix(),
+    )
 
     # Export
     if args.no_export:
         logger.info("Skipping the exporting step...")
     else:
-        export_application(slicer_build_file_path)
+        export_application(args, slicer_build_file_path)
 
 
-def find_geoslicer_base_application_directory_path(slicer_build_file_path: Path) -> Path:
+def find_geoslicer_base_application_directory_path(
+    slicer_build_file_path: Path,
+) -> Path:
     CPack_directory_path = slicer_build_file_path / "_CPack_Packages"
     geoslicer_base_application_directory_path = None
     geoslicer_dir_tag = "GeoSlicer"
@@ -118,7 +143,9 @@ def check_oci_configuration(config, logger=logging):
     logger.info("OCI credentials are okay!")
 
 
-def upload_file_2_bucket(input_file_path, bucket_output_directory, namespace, bucket_name):
+def upload_file_2_bucket(
+    input_file_path, bucket_output_directory, namespace, bucket_name
+):
     logger.info("Uploading file to OCI bucket...")
     config = oci.config.from_file()
     check_oci_configuration(config)
@@ -129,15 +156,19 @@ def upload_file_2_bucket(input_file_path, bucket_output_directory, namespace, bu
 
     with open(input_file_path.as_posix(), "rb") as file:
         object_storage_client = oci.object_storage.ObjectStorageClient(config)
-        object_storage_client.put_object(namespace, bucket_name, output_bucket_file_path.as_posix(), file)
+        object_storage_client.put_object(
+            namespace, bucket_name, output_bucket_file_path.as_posix(), file
+        )
 
     logger.info(
         f"Application base compressed file successfully uploaded. Bucket file path: {output_bucket_file_path.as_posix()}"
     )
 
 
-def export_application(slicer_build_file_path: Path):
-    geoslicer_base_directory_path = find_geoslicer_base_application_directory_path(slicer_build_file_path)
+def export_application(args, slicer_build_file_path: Path):
+    geoslicer_base_directory_path = find_geoslicer_base_application_directory_path(
+        slicer_build_file_path
+    )
     if geoslicer_base_directory_path is None:
         raise RuntimeError(
             f"Couldn't find the GeoSlicer base application directory in {slicer_build_file_path.as_posix()}"
@@ -149,7 +180,12 @@ def export_application(slicer_build_file_path: Path):
         geoslicer_base_directory_path.name
     )
     archive_format = "zip" if sys.platform == "win32" else "gztar"
-    shutil.make_archive(geoslicer_base_compressed_file_path, archive_format, geoslicer_base_directory_path.parent, geoslicer_base_directory_path.name)
+    shutil.make_archive(
+        geoslicer_base_compressed_file_path,
+        archive_format,
+        geoslicer_base_directory_path.parent,
+        geoslicer_base_directory_path.name,
+    )
 
     archive_extension = "zip" if sys.platform == "win32" else "tar.gz"
     geoslicer_base_compressed_file_path = geoslicer_base_directory_path.parent / (
@@ -160,15 +196,32 @@ def export_application(slicer_build_file_path: Path):
     )
 
     # Export
-    bucket_output_directory = f"GeoSlicer/base/{sys.platform}"
+    build_type = str(args.type).lower()
+    bucket_output_directory = f"GeoSlicer/base/{build_type}/{sys.platform}"
     bucket_name = "General_ltrace_files"
     namespace = "grrjnyzvhu1t"
-    upload_file_2_bucket(geoslicer_base_compressed_file_path, bucket_output_directory, namespace, bucket_name)
+    upload_file_2_bucket(
+        geoslicer_base_compressed_file_path,
+        bucket_output_directory,
+        namespace,
+        bucket_name,
+    )
 
 
-def run_subprocess(command, assert_exit_code=True, shell=True, cwd=None):
+def run_subprocess(command, assert_exit_code=True, shell=True, cwd=None, env=None):
     """Wrapper for running subprocess and reading its output"""
-    with subprocess.Popen(command, stdout=subprocess.PIPE, bufsize=1, universal_newlines=True, shell=shell, cwd=cwd) as proc:
+    if isinstance(command, str):
+        command = command.split()
+
+    with subprocess.Popen(
+        command,
+        stdout=subprocess.PIPE,
+        bufsize=1,
+        universal_newlines=True,
+        shell=shell,
+        cwd=cwd,
+        env=env,
+    ) as proc:
         for line in proc.stdout:
             print(f"\t{line}", end="")
 
@@ -184,11 +237,12 @@ def onerror(func, path, exc_info):
     it attempts to add write permission and then retries.
 
     If the error is for another reason it re-raises the error.
-    
+
     Usage : ``shutil.rmtree(path, onerror=onerror)``
     Reference: https://stackoverflow.com/questions/2656322/shutil-rmtree-fails-on-windows-with-access-is-denied
     """
     import stat
+
     # Is the error an access error?
     if not os.access(path, os.W_OK):
         os.chmod(path, stat.S_IWUSR)
@@ -197,19 +251,56 @@ def onerror(func, path, exc_info):
         raise
 
 
+def only_export_process(args):
+    output_directory_path = get_working_dir(args)
+    slicer_build_file_path = output_directory_path / "Slicer-build"
+    export_application(args, slicer_build_file_path)
+
+
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="Apply dynamic changes to CMakeLists.txt")
-    parser.add_argument("--source", help="The source code directory path.", default=None)
-    parser.add_argument("--jobs", help="The jobs quantity for parallel building.", default=1)
-    parser.add_argument("--type", help="The build type. Default to Release", default="Release")
-    parser.add_argument("--no-export", action="store_true", help="Skip application exporting step", default=False)
-    parser.add_argument("--no-cache", action="store_true", help="Delete old build files before starting process", default=False)
-    parser.add_argument("--avoid-long-path", action="store_true", help="Avoid long path issues", default=False)
+    parser = argparse.ArgumentParser(
+        description="Apply dynamic changes to CMakeLists.txt"
+    )
+    parser.add_argument(
+        "--source", help="The source code directory path.", default=None
+    )
+    parser.add_argument(
+        "--jobs", help="The jobs quantity for parallel building.", default=1
+    )
+    parser.add_argument(
+        "--type", help="The build type. Default to Release", default="Release"
+    )
+    parser.add_argument(
+        "--no-export",
+        action="store_true",
+        help="Skip application exporting step",
+        default=False,
+    )
+    parser.add_argument(
+        "--no-cache",
+        action="store_true",
+        help="Delete old build files before starting process",
+        default=False,
+    )
+    parser.add_argument(
+        "--avoid-long-path",
+        action="store_true",
+        help="Avoid long path issues",
+        default=False,
+    )
+    parser.add_argument(
+        "--only-export",
+        action="store_true",
+        help="Only export the generated package",
+        default=False,
+    )
 
     args = parser.parse_args()
 
     if args.source is None:
-        raise AttributeError("The source code directory path is missing! Aborting process...")
+        raise AttributeError(
+            "The source code directory path is missing! Aborting process..."
+        )
 
     args.source = Path(args.source).absolute()
 
@@ -221,5 +312,7 @@ if __name__ == "__main__":
         logger.info(f"Error: {error}")
         sys.exit(1)
 
-    logger.info("The process to build and generate the application package finished with success.")
+    logger.info(
+        "The process to build and generate the application package finished with success."
+    )
     sys.exit(0)
